requirements.txt
docling==2.84.0

langchain==1.2.15
langchain-chroma==1.1.0
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1

chromadb==1.5.5
sentence-transformers==5.3.0
huggingface-hub==1.9.0

fastapi==0.135.3
uvicorn==0.43.0
python-multipart==0.0.22

python-dotenv==1.2.2
httpx==0.28.1
Pillow==12.2.0